Tea-9097: Justeringer etter QA
